Prove a top level correctness result (#37)
Refactor extraction into a new Extraction.v file that is sufficiently general to cover our use cases and prove a top level statement about it correct. To do this I had to

* Strengthen the erasure proof from MetaCoq to require only dependencies and not the whole environment to be erased.
* Change trans_env to translate type aliases into constants with boxes as bodies since these can appear in terms (although they should be removed later, we might want to prove that).
* Prove our recursive erasure procedure correct to be able to apply (1)
* Combine all the bits into a top level statement.

The top level statement is saying that any bool (in fact any constructor) that is not erasable will evaluate to the same constructor in the extracted environment.